requirements.txt
requests
pandas
pymongo
python-dotenv
matplotlib